Chore/cleanup (#58)
* Separate tools folder

* Unified CMakeLists of all projects

* Added web builds and diagnostics test to validate builds

* Format, format script and format test

* Trying pre-commit

* Fix format

* Trying to fix builds. Use static SDL3 where possible always.

* Update CMakeLists.txt

* Compilation error on windows

* Fixed stupid log stupid macros stupid conditional compilation

* Trying to make this worl

* Update build-web.yml

* Update build-web.yml

* Web build should copy index.html now

* Update build-web.yml

* Update build-web.yml

* Updated CMake files

* Combined build-web.yml into cmake-multi-platform.yml  using a matrix strategy and updated CMake files to make sure web build generates an index.html

* ci: fix emsdk cache error, windows artifact step, and node warnings

* fix(web): ensure BUILD_WEB scope sets OUTPUT_NAME index and guarantee index.html in web artifact

* fix(web): set target SUFFIX .html to force emcc to generate index.html and add index.html launcher fallback

* feat(web): add customizable src/index.html templates for all examples and tools

* itch.io deploy action with Butler

* Cleaned type warnings

* Improved sample scenes

+++ examples/sample-scenes/include/CollisionHandling.h
@@ -8,7 +8,7 @@ using namespace WeirdEngine;
 class CollisionHandlingScene : public Scene2D
 {
 public:
-	CollisionHandlingScene(){};
+	CollisionHandlingScene() {};
 
 private:
 	// Inherited via Scene
@@ -21,8 +21,8 @@ private:
 
 		for (size_t i = 0; i < 10; i++)
 		{
-			float y = 10 + i;
-			float x = 2 * i;
+			float y = 10.0f + static_cast<float>(i);
+			float x = 2.0f * static_cast<float>(i);
 
 			int material = 4 + (i % 12);
